Moves Msg::preparsecode() to Parser::sanitize()
