+++ Sources/Actions/Admin/News.php
@@ -28,7 +28,6 @@ use SMF\Lang;
 use SMF\Logging;
 use SMF\Mail;
 use SMF\Menu;
-use SMF\Msg;
 use SMF\Parser;
 use SMF\PersonalMessage\PM;
 use SMF\SecurityToken;
@@ -303,12 +302,15 @@ class News implements ActionInterface
 			User::$me->checkSession();
 
 			foreach ($_POST['news'] as $i => $news) {
-				if (trim($news) == '') {
+				$news = Utils::htmlTrim($news);
+
+				if ($news == '') {
 					unset($_POST['news'][$i]);
 				} else {
-					$_POST['news'][$i] = Utils::htmlspecialchars($_POST['news'][$i], ENT_QUOTES);
-
-					Msg::preparsecode($_POST['news'][$i], false, !empty(Config::$modSettings['autoLinkUrls']));
+					$_POST['news'][$i] = Parser::sanitize(
+						Utils::htmlspecialchars($news, ENT_QUOTES),
+						autolink: !empty(Config::$modSettings['autoLinkUrls']),
+					);
 				}
 			}
 
@@ -1200,7 +1202,7 @@ class News implements ActionInterface
 				continue;
 			}
 
-			Msg::preparsecode(Utils::$context[$key], false, !empty(Config::$modSettings['autoLinkUrls']));
+			Utils::$context[$key] = Parser::sanitize(Utils::$context[$key], autolink: !empty(Config::$modSettings['autoLinkUrls']));
 
 			if (!empty(Utils::$context['send_html'])) {
 				$enablePostHTML = Config::$modSettings['enablePostHTML'];
