+++ Sources/Actions/XmlHttp.php
@@ -26,7 +26,6 @@ use SMF\Editor;
 use SMF\ErrorHandler;
 use SMF\IntegrationHook;
 use SMF\Lang;
-use SMF\Msg;
 use SMF\OutputTypeInterface;
 use SMF\OutputTypes;
 use SMF\Parser;
@@ -185,7 +184,7 @@ class XmlHttp implements ActionInterface, Routable
 		if (empty($news)) {
 			$errors[] = ['value' => 'no_news'];
 		} else {
-			Msg::preparsecode($news, false, !empty(Config::$modSettings['autoLinkUrls']));
+			$news = Parser::sanitize($news, autolink: !empty(Config::$modSettings['autoLinkUrls']));
 		}
 
 		Utils::$context['xml_data'] = [
@@ -389,9 +388,7 @@ class XmlHttp implements ActionInterface, Routable
 			}
 
 			if (!empty($_POST['body'])) {
-				Msg::preparsecode($warning_body, false, !empty(Config::$modSettings['autoLinkUrls']));
-
-				$warning_body = Parser::transform($warning_body);
+				$warning_body = Parser::transform(Parser::sanitize($warning_body, autolink: !empty(Config::$modSettings['autoLinkUrls'])));
 				$warning_body = Utils::adjustHeadingLevels($warning_body, null);
 			}
