+++ Sources/Parser.php
@@ -413,6 +413,332 @@ abstract class Parser
 	}
 
 	/**
+	 * Cleans and repairs BBCode in a string of user input.
+	 *
+	 * Formerly known as preparsecode().
+	 *
+	 * @param string $message The message.
+	 * @param bool $previewing Whether we're previewing. Default: false.
+	 * @param bool $autolink Whether to autolink plain-text URLs. Default: false.
+	 * @return string Cleaned version of $message.
+	 */
+	public static function sanitize(string $message, bool $previewing = false, bool $autolink = false): string
+	{
+		static $tags_regex, $disallowed_tags_regex;
+
+		// Convert control characters (except \t, \r, and \n) to harmless Unicode symbols
+		$message = strtr(
+			$message,
+			[
+				"\x00" => '&#x2400;', "\x01" => '&#x2401;', "\x02" => '&#x2402;',
+				"\x03" => '&#x2403;', "\x04" => '&#x2404;', "\x05" => '&#x2405;',
+				"\x06" => '&#x2406;', "\x07" => '&#x2407;', "\x08" => '&#x2408;',
+				"\x0b" => '&#x240b;', "\x0c" => '&#x240c;', "\x0e" => '&#x240e;',
+				"\x0f" => '&#x240f;', "\x10" => '&#x2410;', "\x11" => '&#x2411;',
+				"\x12" => '&#x2412;', "\x13" => '&#x2413;', "\x14" => '&#x2414;',
+				"\x15" => '&#x2415;', "\x16" => '&#x2416;', "\x17" => '&#x2417;',
+				"\x18" => '&#x2418;', "\x19" => '&#x2419;', "\x1a" => '&#x241a;',
+				"\x1b" => '&#x241b;', "\x1c" => '&#x241c;', "\x1d" => '&#x241d;',
+				"\x1e" => '&#x241e;', "\x1f" => '&#x241f;',
+			],
+		);
+
+		// Normalize Unicode characters for storage efficiency, better searching, etc.
+		$message = Utils::normalize($message);
+
+		// Clean out any other funky stuff.
+		$message = Utils::sanitizeChars($message, 0);
+
+		// Clean up after nobbc ;).
+		$message = preg_replace_callback(
+			'~\[nobbc[^\]]*\](.+?)\[/nobbc\]~is',
+			fn($matches) => '[nobbc]' . strtr($matches[1], ['[' => '&#91;', ']' => '&#93;', ':' => '&#58;', '@' => '&#64;']) . '[/nobbc]',
+			$message,
+		);
+
+		// Remove \r's... they're evil!
+		$message = strtr($message, ["\r\n" => "\n", "\r" => "\n"]);
+
+		// You won't believe this - but too many periods upsets apache it seems!
+		// @todo You're right, I don't believe this. Verify or remove.
+		$message = preg_replace('~\.{100,}~', '...', $message);
+
+		// Trim off trailing quotes - these often happen by accident.
+		while (str_ends_with($message, '[quote]')) {
+			$message = substr($message, 0, -7);
+		}
+
+		while (str_starts_with($message, '[/quote]')) {
+			$message = substr($message, 8);
+		}
+
+		if (str_contains($message, '[cowsay') && !User::$me->allowedTo('bbc_cowsay')) {
+			$message = preg_replace('~\[(/?)cowsay[^\]]*\]~iu', '[$1pre]', $message);
+		}
+
+		// Find all code blocks, work out whether we'd be parsing them, then ensure they are all closed.
+		$in_tag = false;
+		$had_tag = false;
+		$codeopen = 0;
+
+		if (preg_match_all('~(\[(/)*code(?:=[^\]]+)?\])~is', $message, $matches)) {
+			foreach ($matches[0] as $index => $dummy) {
+				// Closing?
+				if (!empty($matches[2][$index])) {
+					// If it's closing and we're not in a tag we need to open it...
+					if (!$in_tag) {
+						$codeopen = true;
+					}
+
+					// Either way we ain't in one any more.
+					$in_tag = false;
+				}
+				// Opening tag...
+				else {
+					$had_tag = true;
+
+					// If we're in a tag don't do nought!
+					if (!$in_tag) {
+						$in_tag = true;
+					}
+				}
+			}
+		}
+
+		// If we have an open tag, close it.
+		if ($in_tag) {
+			$message .= '[/code]';
+		}
+
+		// Open any ones that need to be open, only if we've never had a tag.
+		if ($codeopen && !$had_tag) {
+			$message = '[code]' . $message;
+		}
+
+		// Replace code BBC with placeholders. We'll restore them at the end.
+		$parts = preg_split('/(\[code(?:=[^\]]+)?\](?:[^\[]|\[(?!\/code\])|(?R))*\[\/code])/i', $message, -1, PREG_SPLIT_DELIM_CAPTURE);
+
+		for ($i = 0, $n = \count($parts); $i < $n; $i++) {
+			if ($i % 2 == 1) {
+				$substitute = md5($parts[$i]);
+				$code_tags[$substitute] = $parts[$i];
+				$parts[$i] = $substitute;
+			}
+		}
+
+		$message = implode('', $parts);
+
+		// Autolink any plain-text URLs.
+		if (!empty($autolink)) {
+			$message = Autolinker::load()->makeLinks($message);
+		}
+
+		// Now let's fix the img and url tags.
+		$message = Autolinker::load()->fixUrlsInBBC($message);
+		$message = self::fixTags($message);
+
+		// Replace /me.+?\n with [me=name]dsf[/me]\n.
+		if (
+			str_contains(User::$me->name, '[')
+			|| str_contains(User::$me->name, ']')
+			|| str_contains(User::$me->name, '\'')
+			|| str_contains(User::$me->name, '"')
+		) {
+			$message = preg_replace(
+				'~(\A|\n)/me(?: |&nbsp;)([^\n]*)(?:\z)?~i',
+				'$1[me=&quot;' . User::$me->name . '&quot;]$2[/me]',
+				$message,
+			);
+		} else {
+			$message = preg_replace(
+				'~(\A|\n)/me(?: |&nbsp;)([^\n]*)(?:\z)?~i',
+				'$1[me=' . User::$me->name . ']$2[/me]',
+				$message,
+			);
+		}
+
+		// Clean up the [html] tags and content.
+		if (!$previewing && str_contains($message, '[html]')) {
+			if (User::$me->allowedTo('bbc_html')) {
+				$message = preg_replace_callback(
+					'~\[html\](.+?)\[/html\]~is',
+					fn($matches) => '[html]' . strtr(Utils::htmlspecialcharsDecode($matches[1]), ["\n" => '&#13;', '  ' => ' &#32;', '[' => '&#91;', ']' => '&#93;']) . '[/html]',
+					$message,
+				);
+			}
+			// We should edit them out, or else if an admin edits the message they will get shown...
+			else {
+				while (str_contains($message, '[html]')) {
+					$message = preg_replace('~\[/?html\]~i', '', $message);
+				}
+			}
+		}
+
+		// Let's look at the time tags...
+		$message = preg_replace_callback(
+			'~\[time(?:=([^\\]]*))?\](.+?)\[/time\]~i',
+			function ($matches) {
+				return preg_replace(
+					[
+						'~^<time[^>]*\bdatetime="([^"]+)"[^>]*>(.*)</time>$~',
+						'~^<span[^>]*>.*</span>$~',
+					],
+					[
+						// If it parsed successfully, insert the resolved datetime value.
+						// This ensures that "[time]today[/time]" ends up resolving to
+						// the date the post was written, not the date it is being read.
+						'[time=$1]$2[/time]',
+						// If it didn't parse successfully, remove the BBC entirely.
+						$matches[2],
+					],
+					self::transform($matches[0], self::INPUT_BBC),
+				);
+			},
+			$message,
+		);
+
+		// Change the color specific tags to [color=the color].
+		// First do the opening tags.
+		$message = preg_replace('~\[(black|blue|green|red|white)\]~', '[color=$1]', $message);
+
+		// And now do the closing tags
+		$message = preg_replace('~\[/(black|blue|green|red|white)\]~', '[/color]', $message);
+
+		// Neutralize any BBC tags this member isn't permitted to use.
+		if (empty($disallowed_tags_regex)) {
+			// Legacy BBC are only retained for historical reasons.
+			// They're not for use in new posts.
+			$disallowed_bbc = Utils::$context['legacy_bbc'];
+
+			// Some BBC require permissions.
+			foreach (Utils::$context['restricted_bbc'] as $bbc) {
+				// Skip html, since we handled it separately above.
+				if ($bbc === 'html') {
+					continue;
+				}
+
+				if (!User::$me->allowedTo('bbc_' . $bbc)) {
+					$disallowed_bbc[] = $bbc;
+				}
+			}
+
+			$disallowed_tags_regex = Utils::buildRegex(array_unique($disallowed_bbc), '~');
+		}
+
+		if (!empty($disallowed_tags_regex)) {
+			$message = preg_replace('~\[(?=/?' . $disallowed_tags_regex . '\b)~i', '&#91;', $message);
+		}
+
+		// Make sure all tags are lowercase.
+		$message = preg_replace_callback(
+			'~\[(/?)(list|li|table|tr|td)\b([^\]]*)\]~i',
+			fn($m) => '[' . $m[1] . strtolower($m[2]) . $m[3] . ']',
+			$message,
+		);
+
+		$list_open = substr_count($message, '[list]') + substr_count($message, '[list ');
+		$list_close = substr_count($message, '[/list]');
+
+		if ($list_close - $list_open > 0) {
+			$message = str_repeat('[list]', $list_close - $list_open) . $message;
+		}
+
+		if ($list_open - $list_close > 0) {
+			$message = $message . str_repeat('[/list]', $list_open - $list_close);
+		}
+
+		$mistake_fixes = [
+			// Find [table]s not followed by [tr].
+			'~\[table\](?![\s\x{A0}]*\[tr\])~su' => '[table][tr]',
+			// Find [tr]s not followed by [td].
+			'~\[tr\](?![\s\x{A0}]*\[td\])~su' => '[tr][td]',
+			// Find [/td]s not followed by something valid.
+			'~\[/td\](?![\s\x{A0}]*(?:\[td\]|\[/tr\]|\[/table\]))~su' => '[/td][/tr]',
+			// Find [/tr]s not followed by something valid.
+			'~\[/tr\](?![\s\x{A0}]*(?:\[tr\]|\[/table\]))~su' => '[/tr][/table]',
+			// Find [/td]s incorrectly followed by [/table].
+			'~\[/td\][\s\x{A0}]*\[/table\]~su' => '[/td][/tr][/table]',
+			// Find [table]s, [tr]s, and [/td]s (possibly correctly) followed by [td].
+			'~\[(table|tr|/td)\]([\s\x{A0}]*)\[td\]~su' => '[$1]$2[_td_]',
+			// Now, any [td]s left should have a [tr] before them.
+			'~\[td\]~s' => '[tr][td]',
+			// Look for [tr]s which are correctly placed.
+			'~\[(table|/tr)\]([\s\x{A0}]*)\[tr\]~su' => '[$1]$2[_tr_]',
+			// Any remaining [tr]s should have a [table] before them.
+			'~\[tr\]~s' => '[table][tr]',
+			// Look for [/td]s followed by [/tr].
+			'~\[/td\]([\s\x{A0}]*)\[/tr\]~su' => '[/td]$1[_/tr_]',
+			// Any remaining [/tr]s should have a [/td].
+			'~\[/tr\]~s' => '[/td][/tr]',
+			// Look for properly opened [li]s which aren't closed.
+			'~\[li\]([^\[\]]+?)\[li\]~s' => '[li]$1[_/li_][_li_]',
+			'~\[li\]([^\[\]]+?)\[/list\]~s' => '[_li_]$1[_/li_][/list]',
+			'~\[li\]([^\[\]]+?)$~s' => '[li]$1[/li]',
+			// Lists - find correctly closed items/lists.
+			'~\[/li\]([\s\x{A0}]*)\[/list\]~su' => '[_/li_]$1[/list]',
+			// Find list items closed and then opened.
+			'~\[/li\]([\s\x{A0}]*)\[li\]~su' => '[_/li_]$1[_li_]',
+			// Now, find any [list]s or [/li]s followed by [li].
+			'~\[(list(?: [^\]]*?)?|/li)\]([\s\x{A0}]*)\[li\]~su' => '[$1]$2[_li_]',
+			// Allow for sub lists.
+			'~\[/li\]([\s\x{A0}]*)\[list\]~u' => '[_/li_]$1[list]',
+			'~\[/list\]([\s\x{A0}]*)\[li\]~u' => '[/list]$1[_li_]',
+			// Any remaining [li]s weren't inside a [list].
+			'~\[li\]~' => '[list][li]',
+			// Any remaining [/li]s weren't before a [/list].
+			'~\[/li\]~' => '[/li][/list]',
+			// Put the correct ones back how we found them.
+			'~\[_(li|/li|td|tr|/tr)_\]~' => '[$1]',
+			// Images with no real url.
+			'~\[img\]https?://.{0,7}\[/img\]~' => '',
+		];
+
+		// Fix up some use of tables without [tr]s, etc. (it has to be done more than once to catch it all.)
+		for ($j = 0; $j < 3; $j++) {
+			$message = preg_replace(array_keys($mistake_fixes), $mistake_fixes, $message);
+		}
+
+		// Remove empty bbc from the sections outside the code tags
+		if (empty($tags_regex)) {
+			$allowed_empty = ['anchor', 'td'];
+
+			$tags = [];
+
+			foreach (self::getBBCodes() as $code) {
+				if (!\in_array($code['tag'], $allowed_empty)) {
+					$tags[] = $code['tag'];
+				}
+			}
+
+			$tags_regex = Utils::buildRegex($tags, '~');
+		}
+
+		while (preg_match('~\[(' . $tags_regex . ')\b[^\]]*\]\s*\[/\1\]\s?~i', $message)) {
+			$message = preg_replace('~\[(' . $tags_regex . ')[^\]]*\]\s*\[/\1\]\s?~i', '', $message);
+		}
+
+		// Restore code blocks
+		if (!empty($code_tags)) {
+			$message = str_replace(array_keys($code_tags), array_values($code_tags), $message);
+		}
+
+		// Restore white space entities
+		if (!$previewing) {
+			$message = strtr($message, ['  ' => '&nbsp; ', "\n" => '<br>', "\u{A0}" => '&nbsp;']);
+		} else {
+			$message = strtr($message, ['  ' => '&nbsp; ', "\u{A0}" => '&nbsp;']);
+		}
+
+		// Now let's quickly clean up things that will slow our parser (which are common in posted code.)
+		$message = strtr($message, ['[]' => '&#91;]', '[&#039;' => '&#91;&#039;']);
+
+		// Any hooks want to work here?
+		IntegrationHook::call('integrate_preparsecode', [&$message, $previewing]);
+
+		return $message;
+	}
+
+	/**
 	 * Given a string that has previously been cleaned by Parser::sanitize(),
 	 * this returns a version suitable for editing in an HTML form.
 	 *
@@ -800,5 +1126,214 @@ abstract class Parser
 			// Additional stuff that might affect output.
 			$cache_key_extras,
 		]));
+	}
+
+	/**
+	 * Helper for Parser::sanitize() that fixes URIs and BBCode markup
+	 * in various BBCode tags.
+	 *
+	 * Ensures URIs are valid, ensures BBCode markup is well-formed, and ensures
+	 * certain security problems are prevented.
+	 *
+	 * @param string $message The message.
+	 * @return string Fixed version of $message.
+	 */
+	protected static function fixTags(string $message): string
+	{
+		// WARNING: Editing the below can cause large security holes in your forum.
+		// Edit only if you are sure you know what you are doing.
+
+		$fixes = [
+			// [img]http://...[/img] or [img width=1]http://...[/img]
+			[
+				'tag' => 'img',
+				'protocols' => ['http', 'https'],
+				'embedded_url' => false,
+				'has_equal_sign' => false,
+				'has_extra' => true,
+			],
+			// [url]http://...[/url]
+			[
+				'tag' => 'url',
+				'protocols' => ['http', 'https'],
+				'embedded_url' => false,
+				'has_equal_sign' => false,
+				'has_extra' => false,
+			],
+			// [url=http://...]name[/url]
+			[
+				'tag' => 'url',
+				'protocols' => ['http', 'https'],
+				'embedded_url' => true,
+				'has_equal_sign' => true,
+				'has_extra' => false,
+			],
+			// [iurl]http://...[/iurl]
+			[
+				'tag' => 'iurl',
+				'protocols' => ['http', 'https'],
+				'embedded_url' => false,
+				'has_equal_sign' => false,
+				'has_extra' => false,
+			],
+			// [iurl=http://...]name[/iurl]
+			[
+				'tag' => 'iurl',
+				'protocols' => ['http', 'https'],
+				'embedded_url' => true,
+				'has_equal_sign' => true,
+				'has_extra' => false,
+			],
+			// The rest of these are deprecated.
+			// [ftp]ftp://...[/ftp]
+			[
+				'tag' => 'ftp',
+				'protocols' => ['ftp', 'ftps', 'sftp'],
+				'embedded_url' => false,
+				'has_equal_sign' => false,
+				'has_extra' => false,
+			],
+			// [ftp=ftp://...]name[/ftp]
+			[
+				'tag' => 'ftp',
+				'protocols' => ['ftp', 'ftps', 'sftp'],
+				'embedded_url' => true,
+				'has_equal_sign' => true,
+				'has_extra' => false,
+			],
+			// [flash]http://...[/flash]
+			[
+				'tag' => 'flash',
+				'protocols' => ['http', 'https'],
+				'embedded_url' => false,
+				'has_equal_sign' => false,
+				'has_extra' => true,
+			],
+		];
+
+		// Fix each type of tag.
+		foreach ($fixes as $fix_info) {
+			$message = self::fixTag($message, ...$fix_info);
+		}
+
+		// Now fix possible security problems with images loading links automatically...
+		$message = preg_replace_callback(
+			'~(\[img.*?\])(.+?)\[/img\]~is',
+			function ($m) {
+				return $m[1] . preg_replace('~action(=|%3d)(?!dlattach)~i', 'action-', $m[2]) . '[/img]';
+			},
+			$message,
+		);
+
+		return $message;
+	}
+
+	/**
+	 * Helper for Parser::fixTags() that fixes URIs and BBCode markup for a
+	 * specific type of BBCode tag.
+	 *
+	 * Ensures URIs are valid, ensures BBCode markup is well-formed, and ensures
+	 * certain security problems are prevented.
+	 *
+	 * @param string $message The message
+	 * @param string $tag The tag
+	 * @param array $protocols The protocols
+	 * @param bool $embedded_url Whether it *can* be set to something
+	 * @param bool $has_equal_sign Whether it *is* set to something
+	 * @param bool $has_extra Whether it can have extra cruft after the begin tag.
+	 * @return string Fixed version of $message.
+	 */
+	protected static function fixTag(string $message, string $tag, array $protocols, bool $embedded_url, bool $has_equal_sign, bool $has_extra): string
+	{
+		$forbidden_protocols = [
+			// Poses security risks.
+			'javascript',
+			// Allows file data to be embedded, bypassing our attachment system.
+			'data',
+		];
+
+		if (preg_match('~^([^:]+://[^/]+)~', Config::$boardurl, $match) != 0) {
+			$domain_url = $match[1];
+		} else {
+			$domain_url = Config::$boardurl . '/';
+		}
+
+		$replaces = [];
+
+		if ($has_equal_sign && $embedded_url) {
+			$quoted = preg_match('~\[(' . $tag . ')=&quot;~', $message);
+
+			preg_match_all('~\[(' . $tag . ')=' . ($quoted ? '&quot;(.*?)&quot;' : '([^\]]*?)') . '\](?:(.+?)\[/(' . $tag . ')\])?~is', $message, $matches);
+		} elseif ($has_equal_sign) {
+			preg_match_all('~\[(' . $tag . ')=([^\]]*?)\](?:(.+?)\[/(' . $tag . ')\])?~is', $message, $matches);
+		} else {
+			preg_match_all('~\[(' . $tag . ($has_extra ? '(?:[^\]]*?)' : '') . ')\](.+?)\[/(' . $tag . ')\]~is', $message, $matches);
+		}
+
+		foreach ($matches[0] as $k => $dummy) {
+			// Remove all leading and trailing whitespace.
+			$replace = trim($matches[2][$k]);
+			$this_tag = $matches[1][$k];
+			$this_close = $has_equal_sign ? (empty($matches[4][$k]) ? '' : $matches[4][$k]) : $matches[3][$k];
+
+			$found = false;
+
+			foreach ($protocols as $protocol) {
+				$found = strncasecmp($replace, $protocol . '://', \strlen($protocol) + 3) === 0;
+
+				if ($found) {
+					break;
+				}
+			}
+
+			$current_protocol = strtolower(Url::create($replace)->scheme ?? '');
+
+			if (\in_array($current_protocol, $forbidden_protocols)) {
+				$replace = 'about:invalid';
+			} elseif (!$found && $protocols[0] == 'http') {
+				// A path
+				if (str_starts_with($replace, '/') && !str_starts_with($replace, '//')) {
+					$replace = $domain_url . $replace;
+				}
+				// A query
+				elseif (str_starts_with($replace, '?')) {
+					$replace = Config::$scripturl . $replace;
+				}
+				// A fragment
+				elseif (str_starts_with($replace, '#') && $embedded_url) {
+					$replace = '#' . preg_replace('~[^A-Za-z0-9_\-#]~', '', substr($replace, 1));
+					$this_tag = 'iurl';
+					$this_close = 'iurl';
+				} elseif (!str_starts_with($replace, '//') && empty($current_protocol)) {
+					$replace = $protocols[0] . '://' . $replace;
+				}
+			} elseif (!$found && $protocols[0] == 'ftp') {
+				$replace = $protocols[0] . '://' . preg_replace('~^(?!ftps?)[^:]+://~', '', $replace);
+			} elseif (!$found && empty($current_protocol)) {
+				$replace = $protocols[0] . '://' . $replace;
+			}
+
+			if ($has_equal_sign && $embedded_url) {
+				$replaces[$matches[0][$k]] = '[' . $this_tag . '=&quot;' . $replace . '&quot;]' . (empty($matches[4][$k]) ? '' : $matches[3][$k] . '[/' . $this_close . ']');
+			} elseif ($has_equal_sign) {
+				$replaces['[' . $matches[1][$k] . '=' . $matches[2][$k] . ']'] = '[' . $this_tag . '=' . $replace . ']';
+			} elseif ($embedded_url) {
+				$replaces['[' . $matches[1][$k] . ']' . $matches[2][$k] . '[/' . $matches[3][$k] . ']'] = '[' . $this_tag . '=' . $replace . ']' . $matches[2][$k] . '[/' . $this_close . ']';
+			} else {
+				$replaces['[' . $matches[1][$k] . ']' . $matches[2][$k] . '[/' . $matches[3][$k] . ']'] = '[' . $this_tag . ']' . $replace . '[/' . $this_close . ']';
+			}
+		}
+
+		foreach ($replaces as $k => $v) {
+			if ($k == $v) {
+				unset($replaces[$k]);
+			}
+		}
+
+		if (!empty($replaces)) {
+			$message = strtr($message, $replaces);
+		}
+
+		return $message;
 	}
 }
